Make VideoGrid Component Reusable
Wanted to re-use the videogrid component I created for the home page on the channel page. So I changed the classNames from hardcoded to variables. I'll use react props to change the variables. With this in mind I had to do some further reorganizing of the project files, and updating the sass.