Set Channel Avatar/Username in Activity Feed
Lifted state up to the render channel page component so above the fold component and activity feed could share username and avatar prop. This allowed me to pull the extraDataFromUrl function out and send that data from the url into the activity feed--in this case the avatar gotten from the url, which is now set to every avatar in the activity feed when a user is on the channel page. This also helped get rid of some unncessary code in the AboveFold component.