Add Clicking Like & Dislike Functionality 🌱 🚧
Also took this opportunity to refactor most of the videoPage code, since the mapping function was unnecessary. Previously it would map the response data to JSX, then save it in state, then render it in the return statement, from state. Now all the JSX starts in the return statement, and simply pulls api data from state. This caused unnecessary re-renders of the component though, so I fixed that by splitting the useEffect hook.